Skip golang-ci lint unused false positives
This commit skips unused violations for specific parts of the code where
it'll be either be used indirectly or unsure about whether it may impact
later on.
