+++ pkg/client/condition.go
@@ -116,9 +116,3 @@ func (cs *conditions) entries() []v1.ClusterOperatorStatusCondition {
 	})
 	return res
 }
-
-type byType []v1.ClusterOperatorStatusCondition
-
-func (b byType) Len() int           { return len(b) }
-func (b byType) Swap(i, j int)      { b[i], b[j] = b[j], b[i] }
-func (b byType) Less(i, j int) bool { return b[i].Type < b[j].Type }
